refactor: 미사용 클래스 삭제
